Organize provisioning file rule structs into new package
This is to keep structs with potential to drift closer together. Eventually, other alerting
file structs should also move into this new package, but the rest require some refactoring
that is out of scope for this PR.
