+++ pkg/services/provisioning/alerting/types.go
@@ -3,6 +3,7 @@ package alerting
 import (
 	"fmt"
 
+	"github.com/grafana/grafana/pkg/services/provisioning/alerting/file"
 	"github.com/grafana/grafana/pkg/services/provisioning/values"
 )
 
@@ -15,8 +16,8 @@ type OrgID int64
 type AlertingFile struct {
 	configVersion
 	Filename            string
-	Groups              []AlertRuleGroup
-	DeleteRules         []RuleDelete
+	Groups              []file.AlertRuleGroupWithFolderTitle
+	DeleteRules         []file.RuleDelete
 	ContactPoints       []ContactPoint
 	DeleteContactPoints []DeleteContactPoint
 	Policies            []NotificiationPolicy
@@ -30,8 +31,8 @@ type AlertingFile struct {
 type AlertingFileV1 struct {
 	configVersion
 	Filename            string
-	Groups              []AlertRuleGroupV1      `json:"groups" yaml:"groups"`
-	DeleteRules         []RuleDeleteV1          `json:"deleteRules" yaml:"deleteRules"`
+	Groups              []file.AlertRuleGroupV1 `json:"groups" yaml:"groups"`
+	DeleteRules         []file.RuleDeleteV1     `json:"deleteRules" yaml:"deleteRules"`
 	ContactPoints       []ContactPointV1        `json:"contactPoints" yaml:"contactPoints"`
 	DeleteContactPoints []DeleteContactPointV1  `json:"deleteContactPoints" yaml:"deleteContactPoints"`
 	Policies            []NotificiationPolicyV1 `json:"policies" yaml:"policies"`
@@ -132,7 +133,7 @@ func (fileV1 *AlertingFileV1) mapRules(alertingFile *AlertingFile) error {
 		if orgID < 1 {
 			orgID = 1
 		}
-		ruleDelete := RuleDelete{
+		ruleDelete := file.RuleDelete{
 			UID:   ruleDeleteV1.UID.Value(),
 			OrgID: orgID,
 		}
